moving public inner classes to own class files
